docs: swipeableNotification 컴포넌트 이동
