+++ src/components/notificationButton/notification.tsx
@@ -4,11 +4,11 @@ import { clsx } from 'clsx';
 import { X } from 'lucide-react';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
-import { SwipeableNotification } from '@/contexts/NotificationDnd';
 import { useDeleteNotification } from '@/hooks/useMutate/useDeleteNotification';
 import { useReadNotification } from '@/hooks/useMutate/useReadNotification';
 import { NotificationType } from '@/types/notification';
 import { getTimeline, parseKST } from '@/utils/date';
+import { SwipeableNotification } from './swipeableNotification';
 
 interface NotificationProps {
   notification: NotificationType;
