+++ src/components/notificationButton/swipeableNotification.tsx
@@ -1,0 +1,33 @@
+'use client';
+
+import { ReactNode } from 'react';
+import { useDraggable } from '@dnd-kit/core';
+import { CSS } from '@dnd-kit/utilities';
+
+interface SwipeableNotificationProps {
+  id: number[];
+  children: ReactNode;
+}
+
+export const SwipeableNotification = ({ id, children }: SwipeableNotificationProps) => {
+  const { setNodeRef, attributes, listeners, transform, isDragging } = useDraggable({
+    id: JSON.stringify(id),
+  });
+
+  const style = {
+    transform: CSS.Transform.toString({
+      x: transform?.x ?? 0,
+      y: 0,
+      scaleX: transform?.scaleX ?? 1,
+      scaleY: transform?.scaleY ?? 1,
+    }),
+    transition: isDragging ? 'none' : 'transform 0.2s ease',
+    opacity: transform?.x ? 1 - Math.abs(transform.x) / 200 : 1,
+  };
+
+  return (
+    <div ref={setNodeRef} {...attributes} {...listeners} className="w-full" style={style}>
+      {children}
+    </div>
+  );
+};
